+++ VaavudElectronicSDK/VEAudioVaavudElectronicDetection.m
@@ -425,7 +425,7 @@ withNumberOfChannels:(UInt32)numberOfChannels {
 
 - (AudioStreamBasicDescription) getAudioStreamBasicDiscriptionOutput {
     
-    size_t bytesPerSample = sizeof (float);
+    UInt32 bytesPerSample = sizeof (float);
     AudioStreamBasicDescription stereoStreamFormat = {0};
     
     
@@ -446,7 +446,7 @@ withNumberOfChannels:(UInt32)numberOfChannels {
 
 - (AudioStreamBasicDescription) getAudioStreamBasicDiscriptionMicrophone {
     
-    size_t bytesPerSample = sizeof (float);
+    UInt32 bytesPerSample = sizeof (float);
     AudioStreamBasicDescription stereoStreamFormat = {0};
